+++ resolwe/test/__init__.py
@@ -15,13 +15,13 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 from resolwe.test.testcases import TestCase, TransactionTestCase
 from resolwe.test.testcases.api import ResolweAPITestCase
 from resolwe.test.testcases.elastic import ElasticSearchTestCase, TransactionElasticSearchTestCase
-from resolwe.test.testcases.process import ProcessTestCase
+from resolwe.test.testcases.process import ProcessTestCase, TransactionProcessTestCase
 from resolwe.test.utils import check_docker, check_installed, with_docker_executor
 
 __all__ = (
     'TestCase', 'TransactionTestCase',
     'ResolweAPITestCase',
     'ElasticSearchTestCase', 'TransactionElasticSearchTestCase',
-    'ProcessTestCase',
+    'ProcessTestCase', 'TransactionProcessTestCase',
     'check_docker', 'check_installed', 'with_docker_executor',
 )